projects/192-simpsons-RAG/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.7
python-dotenv==1.0.0
chromadb==0.4.18